Further conditional cleanup. and refactor to concrete strategy
