+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/odg/figures/ODGGroupFigure.java
@@ -14,7 +14,7 @@ import java.awt.geom.*;
 import java.util.*;
 import org.jhotdraw.draw.*;
 import static org.jhotdraw.draw.AttributeKeys.TRANSFORM;
-import org.jhotdraw.draw.figure.OpacityStrategy;
+import org.jhotdraw.draw.figure.SVGOpacityStrategy;
 import org.jhotdraw.draw.handle.Handle;
 import org.jhotdraw.draw.handle.TransformHandleKit;
 import org.jhotdraw.samples.odg.ODGAttributeKeys;
@@ -68,8 +68,8 @@ public class ODGGroupFigure extends GroupFigure implements ODGFigure {
     @Override
     public void draw(Graphics2D g) {
         double opacity = get(OPACITY);
-        OpacityStrategy ops = new OpacityStrategy(this);
-        if (ops.checkOpacity(g, opacity)){
+        SVGOpacityStrategy ops = new SVGOpacityStrategy(this);
+        if (ops.execute(g, opacity)){
             super.draw(g);
         }
     }
